+++ pymince/text.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
+
 """Useful functions for working with strings."""
+
 import collections
 import functools
 import html
@@ -7,6 +9,7 @@ import random
 import re
 import secrets
 import string
+import unicodedata
 import urllib.parse
 
 import pymince.algorithm
@@ -158,6 +161,25 @@ def remove_decimal_zeros(value, decimal_sep=".", min_decimals=None):
         dec_part = f"{dec_part}{dec_diff}"
 
     return f"{int_part}{decimal_sep}{dec_part}" if dec_part else int_part
+
+
+def slugify(value, allow_unicode=False):
+    """
+    Convert to ASCII if 'allow_unicode' is False. Convert spaces or repeated
+    dashes to single dashes. Remove characters that aren't alphanumerics,
+    underscores, or hyphens. Convert to lowercase. Also strip leading and
+    trailing whitespace, dashes, and underscores.
+
+    https://github.com/django/django/blob/main/django/utils/text.py
+    """
+
+    value = str(value)
+    if allow_unicode:
+        value = unicodedata.normalize("NFKC", value)
+    else:
+        value = unicodedata.normalize("NFKD", value).encode("ascii", "ignore").decode("ascii")
+    value = re.sub(r"[^\w\s-]", "", value.lower())
+    return re.sub(r"[-\s]+", "-", value).strip("-_")
 
 
 def is_url(text, schemes=None, hostnames=None):